refactor: update apr fallback & loading
